requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
asgiref==3.9.1
cachetools==4.2.2
certifi==2025.7.14
charset-normalizer==3.4.2
Django==5.1.7
djangorestframework==3.15.2
exceptiongroup==1.3.0
google-auth==2.38.0
google-genai==1.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.4
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tornado==6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
websockets==14.2
